+++ src/tools/integrations/index.ts
@@ -1,0 +1,1 @@
+import "./list-cost-integrations";
